moving system info observer into it's own file to keep main.js focused on electron setup
